feat: 将用户历史记录 ID 从 int 类型改为 bitint类型以支持更大规模操作

+++ src/main/java/org/opencourse/repositories/HistoryRepo.java
@@ -10,5 +10,5 @@ import org.springframework.stereotype.Repository;
  * @author !EEExp3rt
  */
 @Repository
-public interface HistoryRepo extends JpaRepository<History, Integer> {
+public interface HistoryRepo extends JpaRepository<History, Long> {
 }
